+++ components/patina_acpi/Cargo.toml
@@ -21,6 +21,7 @@ zerocopy = { workspace = true }
 zerocopy-derive = { workspace = true }
 
 patina = { workspace = true, features = ["alloc"] }
+patina_test = { workspace = true }
 
 [dev-dependencies]
 mockall = { workspace = true }
