+++ components/patina_adv_logger/Cargo.toml
@@ -17,7 +17,8 @@ required-features = ['std']
 
 [dependencies]
 log = { workspace = true }
-patina = { workspace = true}
+patina = { workspace = true }
+patina_test = { workspace = true }
 r-efi = { workspace = true }
 spin = { workspace = true }
 mu_rust_helpers = { workspace = true }
