+++ components/patina_test/src/__private_api.rs
@@ -14,30 +14,34 @@ use core::marker::PhantomData;
 
 use r_efi::efi::Guid;
 
-use crate::component::{
+use patina::component::{
     MetaData, Storage, UnsafeStorageCell,
     params::{Param, ParamFunction},
 };
 
+use crate::component::Filter;
+
 /// Where all the test cases marked with `#[patina_test]` are collated to.
-#[cfg(feature = "enable_patina_tests")]
+#[cfg(feature = "test-runner")]
 #[linkme::distributed_slice]
 pub static TEST_CASES: [TestCase];
 
-/// returns the test cases to run.
+/// Returns the test cases to run.
 ///
-/// TEST_CASES exists only when the `enable_patina_tests` feature is
+/// Tests are only collected when the `test-runner` feature is
 /// explicitly enabled. This feature is opt-in and explicit because external
 /// consumers of `patina` who do not register at least one test case with
 /// the `#[patina_test]` attribute may encounter a surprising linker crash (not
 /// just a linker failure), due to the testing infrastructure relying on the
 /// `linkme` crate.
+///
+/// If the `test-runner` feature is not enabled, this function will return an empty slice.
 pub fn test_cases() -> &'static [TestCase] {
-    #[cfg(feature = "enable_patina_tests")]
+    #[cfg(feature = "test-runner")]
     {
         &TEST_CASES
     }
-    #[cfg(not(feature = "enable_patina_tests"))]
+    #[cfg(not(feature = "test-runner"))]
     {
         &[]
     }
@@ -66,7 +70,7 @@ pub struct TestCase {
 }
 
 impl TestCase {
-    pub fn should_run(&self, filters: &[super::Filter]) -> bool {
+    pub fn should_run(&self, filters: &[Filter]) -> bool {
         if self.skip {
             return false;
         }
@@ -76,9 +80,9 @@ impl TestCase {
 
         for filter in filters {
             match filter {
-                super::Filter::Exclude(p) if self.name.contains(p) => return false,
-                super::Filter::Exclude(_) => {}
-                super::Filter::Include(p) => {
+                Filter::Exclude(p) if self.name.contains(p) => return false,
+                Filter::Exclude(_) => {}
+                Filter::Include(p) => {
                     has_includes = true;
                     included |= self.name.contains(p);
                 }
@@ -88,7 +92,7 @@ impl TestCase {
         included || !has_includes
     }
 
-    pub fn run(&self, storage: &mut Storage, debug_mode: bool) -> super::Result {
+    pub fn run(&self, storage: &mut Storage, debug_mode: bool) -> crate::error::Result {
         let ret = if debug_mode {
             log::debug!("#### {} Test Output Start ####", self.name);
             let ret = (self.func)(storage);
@@ -164,9 +168,8 @@ where
 #[coverage(off)]
 mod tests {
     use super::*;
-    use crate::component::Storage;
 
-    use super::super::Filter;
+    extern crate std;
 
     #[test]
     fn test_should_run() {
@@ -310,5 +313,13 @@ mod tests {
 
         let result = test_case.run(&mut storage, false);
         std::assert_eq!(result, Err("Failed to install protocol interface"));
+    }
+
+    #[test]
+    fn test_test_with_invalid_param_combination_is_caught() {
+        assert_eq!(
+            crate::component::tests::TEST_CASE_INVALID.run(&mut Storage::new(), false),
+            Err("Test failed to run due to un-retrievable parameters.")
+        );
     }
 }
