+++ components/patina_test/src/component.rs
@@ -1,0 +1,401 @@
+//! Patina Test Test Runner Component
+//!
+//! This module provides the [TestRunner] component, which is responsible for discovering and registering tests marked
+//! with the `#[patina_test]` attribute for execution. See the [TestRunner] documentation for details on how to
+//! configure and use the test runner. Multiple test runners can be registered to run different sets of tests with
+//! different configurations.
+//!
+//! ## License
+//!
+//! Copyright (c) Microsoft Corporation.
+//!
+//! SPDX-License-Identifier: Apache-2.0
+//!
+use crate::{
+    __private_api,
+    alloc::vec::Vec,
+    service::{Recorder, TestRecord},
+};
+
+use patina::component::{Storage, component};
+
+/// A filter to include or exclude test cases whose name contains the pattern.
+///
+/// # Example
+///
+/// ```rust
+/// use patina_test::component::{TestRunner, Filter};
+///
+/// let runner = TestRunner::default()
+///     .with_filter(Filter::include("x64"))
+///     .with_filter(Filter::exclude("aarch64"));
+/// ```
+#[derive(Debug, Clone)]
+pub enum Filter {
+    /// Only run tests whose name contains the pattern.
+    Include(&'static str),
+    /// Do not run tests whose name contains the pattern.
+    Exclude(&'static str),
+}
+
+impl Filter {
+    /// Creates an include filter. Tests whose name contains `pattern` will be included.
+    pub fn include(pattern: &'static str) -> Self {
+        Self::Include(pattern)
+    }
+
+    /// Creates an exclude filter. Tests whose name contains `pattern` will be excluded.
+    pub fn exclude(pattern: &'static str) -> Self {
+        Self::Exclude(pattern)
+    }
+}
+
+/// A component that runs all test cases marked with the `#[patina_test]` attribute when loaded by the DXE core.
+#[derive(Default, Clone)]
+pub struct TestRunner {
+    filters: Vec<Filter>,
+    debug_mode: bool,
+    fail_callback: Option<fn(&'static str, &'static str)>,
+}
+
+#[component]
+impl TestRunner {
+    /// Adds a filter to control which tests are executed.
+    ///
+    /// The filter pattern is matched against the full test name, which includes the module path.
+    /// For example, if a test is defined in `my_crate::tests`, the name would be
+    /// `my_crate::tests::test_case`.
+    ///
+    /// - [`Filter::Include`]: Only run tests whose name contains the pattern. When multiple include
+    ///   filters are specified, a test runs if it matches **any** of them.
+    /// - [`Filter::Exclude`]: Prevents tests whose name contains the pattern from running. Exclude
+    ///   filters take priority over include filters.
+    ///
+    /// This filter is case-sensitive. It can be called multiple times to add multiple filters.
+    pub fn with_filter(mut self, filter: Filter) -> Self {
+        self.filters.push(filter);
+        self
+    }
+
+    /// Any log messages generated by the test case will be logged if this is set to true.
+    ///
+    /// Defaults to false.
+    pub fn debug_mode(mut self, debug_mode: bool) -> Self {
+        self.debug_mode = debug_mode;
+        self
+    }
+
+    /// Attach a callback function that will be called on test failure.
+    ///
+    /// fn(test_name: &'static str, fail_msg: &'static str)
+    pub fn with_callback(mut self, callback: fn(&'static str, &'static str)) -> Self {
+        self.fail_callback = Some(callback);
+        self
+    }
+
+    /// The entry point for the test runner component.
+    #[coverage(off)]
+    fn entry_point(self, storage: &mut Storage) -> patina::error::Result<()> {
+        let test_list: &'static [__private_api::TestCase] = __private_api::test_cases();
+        self.register_tests(test_list, storage)
+    }
+
+    /// Registers the tests to be executed by the test runner.
+    fn register_tests(
+        &self,
+        test_list: &'static [__private_api::TestCase],
+        storage: &mut Storage,
+    ) -> patina::error::Result<()> {
+        let recorder = match storage.get_service::<Recorder>() {
+            Some(recorder) => recorder,
+            None => {
+                let recorder = Recorder::default();
+                recorder.initialize(storage)?;
+                storage.add_service(recorder);
+                storage.get_service::<Recorder>().expect("Recorder service should be registered.")
+            }
+        };
+
+        let records = test_list
+            .iter()
+            .filter(|&test_case| test_case.should_run(self.filters.as_slice()))
+            .map(|test_case| TestRecord::new(self.debug_mode, test_case, self.fail_callback));
+
+        for record in records {
+            // Only schedule a run if we have not already scheduled for this test.
+            if !recorder.test_registered(record.name()) {
+                record.schedule_run(storage)?;
+            }
+
+            recorder.update_record(record);
+        }
+
+        Ok(())
+    }
+}
+
+#[cfg(test)]
+#[coverage(off)]
+pub(crate) mod tests {
+    extern crate std;
+
+    use super::*;
+    use crate::{
+        __private_api::TestCase,
+        alloc::{boxed::Box, format},
+    };
+    use core::mem::MaybeUninit;
+    use patina::{
+        boot_services::StandardBootServices,
+        component::{IntoComponent, Storage, params::Config},
+    };
+    use r_efi::efi::Guid;
+
+    // A test function where we mock DxeComponentInterface to return what we want for the test.
+    fn test_function(config: Config<i32>) -> crate::error::Result {
+        assert!(*config == 1);
+        Ok(())
+    }
+
+    fn test_function_fail() -> crate::error::Result {
+        Err("Intentional Failure")
+    }
+
+    fn test_function_invalid(_: &mut Storage, _: &mut Storage) -> crate::error::Result {
+        Ok(())
+    }
+
+    // This is mirroring the logic in __private_api.rs to ensure we do properly register test cases.
+    #[linkme::distributed_slice]
+    pub static TEST_TESTS: [super::__private_api::TestCase];
+
+    #[linkme::distributed_slice(TEST_TESTS)]
+    pub static TEST_CASE1: super::__private_api::TestCase = super::__private_api::TestCase {
+        name: "test",
+        triggers: &[super::__private_api::TestTrigger::Manual],
+        skip: false,
+        should_fail: false,
+        fail_msg: None,
+        func: |storage| crate::__private_api::FunctionTest::new(test_function).run(storage.into()),
+    };
+
+    #[linkme::distributed_slice(TEST_TESTS)]
+    pub static TEST_CASE2: super::__private_api::TestCase = super::__private_api::TestCase {
+        name: "test",
+        triggers: &[super::__private_api::TestTrigger::Manual],
+        skip: true,
+        should_fail: false,
+        fail_msg: None,
+        func: |storage| crate::__private_api::FunctionTest::new(test_function).run(storage.into()),
+    };
+
+    pub static TEST_CASE3: super::__private_api::TestCase = super::__private_api::TestCase {
+        name: "test_that_fails",
+        triggers: &[super::__private_api::TestTrigger::Manual],
+        skip: false,
+        should_fail: false,
+        fail_msg: None,
+        func: |storage| crate::__private_api::FunctionTest::new(test_function_fail).run(storage.into()),
+    };
+
+    pub static TEST_CASE4: super::__private_api::TestCase = super::__private_api::TestCase {
+        name: "event_triggered_test",
+        triggers: &[super::__private_api::TestTrigger::Event(&Guid::from_bytes(&[0; 16]))],
+        skip: false,
+        should_fail: false,
+        fail_msg: None,
+        func: |storage| crate::__private_api::FunctionTest::new(test_function_fail).run(storage.into()),
+    };
+
+    pub static TEST_CASE5: super::__private_api::TestCase = super::__private_api::TestCase {
+        name: "timer_triggered_test",
+        triggers: &[super::__private_api::TestTrigger::Timer(1_000_000)],
+        skip: false,
+        should_fail: false,
+        fail_msg: None,
+        func: |storage| crate::__private_api::FunctionTest::new(test_function_fail).run(storage.into()),
+    };
+
+    pub static TEST_CASE_INVALID: super::__private_api::TestCase = super::__private_api::TestCase {
+        name: "invalid_test",
+        triggers: &[super::__private_api::TestTrigger::Event(&Guid::from_bytes(&[0; 16]))],
+        skip: false,
+        should_fail: false,
+        fail_msg: None,
+        func: |storage| crate::__private_api::FunctionTest::new(test_function_invalid).run(storage.into()),
+    };
+
+    #[test]
+    #[ignore = "Skipping test until the service for UEFI services is out, so we can mock it."]
+    fn test_we_can_initialize_the_component() {
+        let mut storage = Storage::new();
+
+        let mut component = super::TestRunner::default().into_component();
+        component.initialize(&mut storage);
+    }
+
+    #[test]
+    #[ignore = "Skipping test until the service for UEFI services is out, so we can mock it."]
+    fn test_we_can_collect_and_execute_tests() {
+        assert_eq!(TEST_TESTS.len(), 2);
+        let mut storage = Storage::new();
+        storage.add_config(1_i32);
+
+        let component = super::TestRunner::default();
+        let result = component.register_tests(&TEST_TESTS, &mut storage);
+        assert!(result.is_ok());
+    }
+
+    #[test]
+    #[ignore = "Skipping test until the service for UEFI services is out, so we can mock it."]
+    fn test_handle_different_test_counts() {
+        let mut storage = Storage::new();
+        storage.add_config(1_i32);
+
+        let test_cases: &'static [TestCase] = Box::leak(Box::new([]));
+        let component = super::TestRunner::default();
+        let result = component.register_tests(test_cases, &mut storage);
+        assert!(result.is_ok());
+
+        let test_cases: &'static [TestCase] = Box::leak(Box::new([TEST_CASE1]));
+        let result = component.register_tests(test_cases, &mut storage);
+        assert!(result.is_ok());
+
+        let test_cases: &'static [TestCase] = Box::leak(Box::new([TEST_CASE1, TEST_CASE2]));
+        let result = component.register_tests(test_cases, &mut storage);
+        assert!(result.is_ok());
+
+        let test_cases: &'static [TestCase] = Box::leak(Box::new([TEST_CASE1, TEST_CASE2, TEST_CASE3]));
+        let result = component.register_tests(test_cases, &mut storage);
+        assert!(result.is_ok());
+    }
+
+    #[test]
+    fn test_func_implements_into_component() {
+        let _ = TestRunner::default().into_component();
+    }
+
+    #[test]
+    fn verify_default_values() {
+        let config = TestRunner::default();
+        assert_eq!(config.filters.len(), 0);
+        assert!(!config.debug_mode);
+    }
+
+    #[test]
+    fn verify_config_sets_properly() {
+        let config = TestRunner::default()
+            .with_filter(Filter::include("aarch64"))
+            .with_filter(Filter::include("test"))
+            .with_filter(Filter::exclude("skip_me"))
+            .debug_mode(true);
+        assert_eq!(config.filters.len(), 3);
+        assert!(config.debug_mode);
+    }
+
+    #[test]
+    #[should_panic(expected = "Callback called")]
+    fn test_test_failure_callback_handler() {
+        let test_runner = crate::component::TestRunner::default().with_callback(|_, _| {
+            panic!("Callback called");
+        });
+
+        let mut storage = Storage::new();
+        storage.add_service(Recorder::default());
+        let bs: MaybeUninit<r_efi::efi::BootServices> = MaybeUninit::uninit();
+
+        // SAFETY: This is very unsafe, because it is not initialized, however this code path only calls create_event
+        // and create_event_ex, which we will fill in with no-op functions.
+        let mut bs = unsafe { bs.assume_init() };
+        extern "efiapi" fn noop_create_event(
+            _type: u32,
+            _tpl: r_efi::efi::Tpl,
+            _notify_function: Option<extern "efiapi" fn(r_efi::efi::Event, *mut core::ffi::c_void)>,
+            _notify_context: *mut core::ffi::c_void,
+            _event: *mut r_efi::efi::Event,
+        ) -> r_efi::efi::Status {
+            r_efi::efi::Status::SUCCESS
+        }
+
+        extern "efiapi" fn noop_create_event_ex(
+            _type: u32,
+            _tpl: r_efi::efi::Tpl,
+            _notify_function: Option<extern "efiapi" fn(r_efi::efi::Event, *mut core::ffi::c_void)>,
+            _notify_context: *const core::ffi::c_void,
+            _guid: *const r_efi::efi::Guid,
+            _event: *mut r_efi::efi::Event,
+        ) -> r_efi::efi::Status {
+            r_efi::efi::Status::SUCCESS
+        }
+
+        bs.create_event = noop_create_event;
+        bs.create_event_ex = noop_create_event_ex;
+
+        storage.set_boot_services(StandardBootServices::new(Box::leak(Box::new(bs))));
+
+        // TEST_CASE3 is designed to fail.
+        let _ = test_runner.register_tests(Box::leak(Box::new([TEST_CASE3])), &mut storage);
+        storage.get_service::<Recorder>().unwrap().run_manual_tests(&mut storage);
+    }
+
+    #[test]
+    fn test_filter_should_work() {
+        let test_runner = TestRunner::default().with_filter(Filter::include("triggered_test"));
+
+        let mut storage = Storage::new();
+        let bs: MaybeUninit<r_efi::efi::BootServices> = MaybeUninit::uninit();
+
+        // SAFETY: This is very unsafe, because it is not initialized, however this code path only calls create_event
+        // create_event_ex, and set_timer which we will fill in with no-op functions.
+        let mut bs = unsafe { bs.assume_init() };
+        extern "efiapi" fn noop_create_event(
+            _type: u32,
+            _tpl: r_efi::efi::Tpl,
+            _notify_function: Option<extern "efiapi" fn(r_efi::efi::Event, *mut core::ffi::c_void)>,
+            _notify_context: *mut core::ffi::c_void,
+            _event: *mut r_efi::efi::Event,
+        ) -> r_efi::efi::Status {
+            r_efi::efi::Status::SUCCESS
+        }
+
+        extern "efiapi" fn noop_create_event_ex(
+            _type: u32,
+            _tpl: r_efi::efi::Tpl,
+            _notify_function: Option<extern "efiapi" fn(r_efi::efi::Event, *mut core::ffi::c_void)>,
+            _notify_context: *const core::ffi::c_void,
+            _guid: *const r_efi::efi::Guid,
+            _event: *mut r_efi::efi::Event,
+        ) -> r_efi::efi::Status {
+            r_efi::efi::Status::SUCCESS
+        }
+
+        extern "efiapi" fn noop_set_timer(
+            _event: r_efi::efi::Event,
+            _type: r_efi::efi::TimerDelay,
+            _trigger_time: u64,
+        ) -> r_efi::efi::Status {
+            r_efi::efi::Status::SUCCESS
+        }
+
+        bs.create_event = noop_create_event;
+        bs.create_event_ex = noop_create_event_ex;
+        bs.set_timer = noop_set_timer;
+
+        storage.set_boot_services(StandardBootServices::new(Box::leak(Box::new(bs))));
+
+        // Failure tests
+        assert!(
+            test_runner.register_tests(Box::leak(Box::new([TEST_CASE3, TEST_CASE4, TEST_CASE5])), &mut storage).is_ok()
+        );
+        let recorder = storage.get_service::<Recorder>().expect("Recorder service should be registered.");
+        recorder.run_manual_tests(&mut storage);
+
+        let output = format!("{}", *recorder);
+
+        // This test is filtered out, so it should not even show up in the results.
+        assert!(!output.contains("test_that_fails"));
+        // This test is not filtered out, but never run, so should log as such.
+        std::println!("{}", output);
+        assert!(output.contains("event_triggered_test ... not triggered"));
+    }
+}
