+++ patina_dxe_core/Cargo.toml
@@ -34,6 +34,7 @@ patina_internal_cpu = { workspace = true }
 patina_debugger = { workspace = true, features = ["alloc"] }
 patina_internal_depex = { workspace = true}
 patina_performance = { workspace = true }
+patina_test = { workspace = true }
 
 [dev-dependencies]
 # To avoid circular dependencies, cargo-release skips dev dependencies when evaluating the release order for
