+++ patina_dxe_core/src/core_patina_tests/stability_tests.rs
@@ -15,16 +15,15 @@ use alloc::vec::Vec;
 use patina::{
     base::{SIZE_1GB, SIZE_2MB, SIZE_4KB},
     pi::dxe_services::GcdMemoryType,
-    test::patina_test,
-    u_assert, u_assert_eq,
 };
 use patina_paging::MemoryAttributes;
+use patina_test::{patina_test, u_assert, u_assert_eq};
 use r_efi::efi;
 
 /// Stability Test: Split a 2MB page into 4KB pages and verify correctness
 #[patina_test]
 #[on(timer = 3_000_000)] // 300ms interval
-fn page_table_tests_2mb_split() -> patina::test::Result {
+fn page_table_tests_2mb_split() -> patina_test::error::Result {
     let mut addr_vec = Vec::new();
     for _ in 0..19 {
         addr_vec.push(
@@ -157,7 +156,7 @@ fn page_table_tests_2mb_split() -> patina::test::Result {
 /// Stability Test: Split a 1GB page into 4KB pages and verify correctness
 #[patina_test]
 #[on(timer = 3_000_000)] // 300ms interval
-fn page_table_tests_1gb_split() -> patina::test::Result {
+fn page_table_tests_1gb_split() -> patina_test::error::Result {
     let addr = GCD.allocate_memory_space(
         AllocateType::TopDown(None),
         GcdMemoryType::SystemMemory,
@@ -242,7 +241,7 @@ fn page_table_tests_1gb_split() -> patina::test::Result {
 /// Stability Test: Map a 2MB page, unmap it, map a 4KB region in it, pattern it, flush tlbs, and verify pattern
 #[patina_test]
 #[on(timer = 3_000_000)] // 300ms interval
-fn page_table_tests_2mb_unmap() -> patina::test::Result {
+fn page_table_tests_2mb_unmap() -> patina_test::error::Result {
     let mut addr_vec = Vec::new();
     for _ in 0..19 {
         addr_vec.push(
@@ -336,7 +335,7 @@ fn page_table_tests_2mb_unmap() -> patina::test::Result {
 /// Stability Test: Map a 1GB page, unmap it, map a 2MB region in it, pattern it, flush tlbs, and verify pattern
 #[patina_test]
 #[on(timer = 3_000_000)] // 300ms interval
-fn page_table_tests_1gb_unmap_2mb_remap() -> patina::test::Result {
+fn page_table_tests_1gb_unmap_2mb_remap() -> patina_test::error::Result {
     let addr = GCD.allocate_memory_space(
         AllocateType::TopDown(None),
         GcdMemoryType::SystemMemory,
@@ -432,7 +431,7 @@ fn page_table_tests_1gb_unmap_2mb_remap() -> patina::test::Result {
 /// Stability Test: Map a 1GB page, unmap it, map a 4KB region in it, pattern it, flush tlbs, and verify pattern
 #[patina_test]
 #[on(timer = 3_000_000)] // 300ms interval
-fn page_table_tests_1gb_unmap_4kb_remap() -> patina::test::Result {
+fn page_table_tests_1gb_unmap_4kb_remap() -> patina_test::error::Result {
     let addr = GCD.allocate_memory_space(
         AllocateType::TopDown(None),
         GcdMemoryType::SystemMemory,
