+++ sdk/patina/src/lib.rs
@@ -55,7 +55,5 @@ pub mod pi;
 pub mod runtime_services;
 pub mod serial;
 #[cfg(any(test, feature = "alloc"))]
-pub mod test;
-#[cfg(any(test, feature = "alloc"))]
 pub mod tpl_mutex;
 pub mod uefi_protocol;
